chore: address review feedback on part delete feature
- Fix Zod 4 schema: replace deprecated required_error with error param
- Extract PartDeleteButton pure logic into shared app/utils/partDeleteHelpers.ts
- Remove unused partStatus prop from PartDeleteButton
- Reframe partIdParamSchema tests to reflect actual validation (non-empty string)
