+++ tests/unit/components/PartDeleteButton.test.ts
@@ -1,42 +1,17 @@
 /**
  * Unit tests for PartDeleteButton pure logic.
  *
- * Tests extracted pure functions that mirror the component's internal logic.
- * No Vue component mounting — follows the same pattern as PartCreationPanel.test.ts.
+ * Tests the shared helper functions that the component imports.
  *
  * Feature: admin-part-delete
  */
 import { describe, it, expect } from 'vitest'
-
-// ---- Pure logic functions extracted from PartDeleteButton.vue ----
-
-/** Determines whether the delete button should be visible (admin-only). */
-export function isDeleteVisible(isAdmin: boolean): boolean {
-  return isAdmin
-}
-
-/** Extracts a user-friendly error message from an API error response. */
-export function extractDeleteError(e: any): string {
-  const status = e?.response?.status ?? e?.statusCode
-  if (status === 403) return 'Admin access required'
-  return e?.data?.message ?? e?.message ?? 'Failed to delete part'
-}
-
-/** Builds the confirmation message for the delete modal. */
-export function buildConfirmationMessage(partId: string): string {
-  return `This will permanently delete part ${partId} and all associated data (certificates, step statuses, overrides). This action cannot be undone.`
-}
-
-/** Builds the success toast after deletion. */
-export function buildDeleteSuccessToast(partId: string): { title: string, description: string, color: string } {
-  return {
-    title: 'Part deleted',
-    description: `${partId} has been permanently removed`,
-    color: 'success',
-  }
-}
-
-// ---- Tests ----
+import {
+  isDeleteVisible,
+  extractDeleteError,
+  buildConfirmationMessage,
+  buildDeleteSuccessToast,
+} from '~/app/utils/partDeleteHelpers'
 
 describe('PartDeleteButton', () => {
   describe('isDeleteVisible', () => {
