+++ tests/unit/schemas/partSchemas.test.ts
@@ -2,13 +2,12 @@ import { describe, it, expect } from 'vitest'
 import { partIdParamSchema } from '../../../server/schemas/partSchemas'
 
 describe('partIdParamSchema', () => {
-  it('accepts a valid part_ prefixed id', () => {
-    const result = partIdParamSchema.safeParse({ id: 'part_00001' })
-    expect(result.success).toBe(true)
-  })
-
-  it('accepts a legacy SN- prefixed id', () => {
-    const result = partIdParamSchema.safeParse({ id: 'SN-00001' })
+  it.each([
+    ['part_00001'],
+    ['SN-00001'],
+    ['any-non-empty-string'],
+  ])('accepts non-empty string: %s', (id) => {
+    const result = partIdParamSchema.safeParse({ id })
     expect(result.success).toBe(true)
   })
